+++ lib/theme_data.dart
@@ -1,19 +1,14 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:google_fonts/google_fonts.dart';
 import 'package:flex_color_scheme/flex_color_scheme.dart';
-
-// Nice to get colors: https://icolorpalette.com/color/101920
-
-// Scheme builder: https://rydmike.com/flexcolorscheme/themesplayground-v8-0/
-
-const Color checkMKIconColor = Color(0xFF13D389);
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 
 /// The [AppTheme] defines light and dark themes for the app.
 ///
 /// Theme setup for FlexColorScheme package v8.
 /// Use same major flex_color_scheme package version. If you use a
 /// lower minor version, some properties may not be supported.
+/// In that case, remove them after copying this theme to your
+/// app or upgrade package to version 8.0.2.
 ///
 /// Use in [MaterialApp] like this:
 ///
@@ -24,54 +19,59 @@ const Color checkMKIconColor = Color(0xFF13D389);
 /// );
 sealed class AppTheme {
   // The defined light theme.
-  static ThemeData lightBase = FlexThemeData.light(
+  static ThemeData light = FlexThemeData.light(
     scheme: FlexScheme.greenM3,
+    fixedColorStyle: FlexFixedColorStyle.seeded,
     subThemesData: const FlexSubThemesData(
       interactionEffects: true,
       tintedDisabledControls: true,
       useM2StyleDividerInM3: true,
+      adaptiveSplash: FlexAdaptive.all(),
       inputDecoratorIsFilled: true,
       inputDecoratorBorderType: FlexInputBorderType.outline,
       alignedDropdown: true,
+      appBarCenterTitle: false,
       navigationRailUseIndicator: true,
       navigationRailLabelType: NavigationRailLabelType.all,
     ),
-    keyColors: const FlexKeyColors(),
     visualDensity: FlexColorScheme.comfortablePlatformDensity,
     cupertinoOverrideTheme: const CupertinoThemeData(applyThemeToAll: true),
   );
-  static ThemeData light = lightBase.copyWith(
-    appBarTheme: AppBarTheme(
-      color: checkMKIconColor,
-      iconTheme: IconThemeData(color: Colors.white),
-    ),
-    textTheme: GoogleFonts.latoTextTheme(lightBase.textTheme),
-  );
-
   // The defined dark theme.
-  static ThemeData darkBase = FlexThemeData.dark(
+  static ThemeData dark = FlexThemeData.dark(
     scheme: FlexScheme.greenM3,
+    appBarOpacity: 0.50,
+    surfaceTint: const Color(0xFF13D389),
     subThemesData: const FlexSubThemesData(
       interactionEffects: true,
       tintedDisabledControls: true,
       blendOnColors: true,
       useM2StyleDividerInM3: true,
+      adaptiveSplash: FlexAdaptive.all(),
       inputDecoratorIsFilled: true,
       inputDecoratorBorderType: FlexInputBorderType.outline,
       alignedDropdown: true,
+      appBarCenterTitle: false,
+      searchBarBackgroundSchemeColor: SchemeColor.onPrimary,
+      searchViewBackgroundSchemeColor: SchemeColor.onPrimary,
       navigationRailUseIndicator: true,
       navigationRailLabelType: NavigationRailLabelType.all,
     ),
-    keyColors: const FlexKeyColors(),
+    keyColors: const FlexKeyColors(
+      useSecondary: true,
+      useTertiary: true,
+      useError: true,
+      keepPrimary: true,
+      keepSecondary: true,
+      keepTertiary: true,
+      keepError: true,
+      keepPrimaryContainer: true,
+      keepSecondaryContainer: true,
+      keepTertiaryContainer: true,
+      keepErrorContainer: true,
+    ),
+    tones: FlexSchemeVariant.material.tones(Brightness.dark),
     visualDensity: FlexColorScheme.comfortablePlatformDensity,
     cupertinoOverrideTheme: const CupertinoThemeData(applyThemeToAll: true),
   );
-  static ThemeData dark = darkBase.copyWith(
-    appBarTheme: AppBarTheme(
-      color: checkMKIconColor,
-      iconTheme: IconThemeData(color: Colors.white),
-    ),
-    textTheme: GoogleFonts.latoTextTheme(darkBase.textTheme),
-  );
-
 }